Add files via upload
The Flexible Bus fare System is a program which serves hands-on booking facility. It includes only Ordinary and Express buses and their  respective fares .The aim of the project is to make ticketing process easy which reduces human involvement. To increase the reach (to every human being at any time) of bus. The Technologies are used are HTML, CSS for web scripting , Python(Django) for developing ,and DbSQlite for Database creation. The Features about the Project are Simple Interface, Fares are displayed according Source and Destination and online ticketing facility.